完善了鉴权的调用，完成了用户中心的接口并测试，完成网关路由到用户中心。

+++ OSCP-SystemAPI/src/main/java/org/joker/oscp/system/api/user/UserApi.java
@@ -38,7 +38,7 @@ public interface UserApi {
      * @param uuid
      * @return 返回 {@link UserInfoModel}
      */
-    UserInfoModel getUserInfo(int uuid);
+    UserInfoModel getUserInfo(Long uuid);
 
     /**
      * 更新用户信息
